asan-visual-ai/backend/requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Settings
pydantic-settings>=2.2.1

# Database
sqlalchemy>=2.0.30
alembic>=1.13.1
asyncpg>=0.29.0

# AI / ML
google-genai>=1.14.0
ultralytics>=8.2.18
torch>=2.9.0
torchvision>=0.22.0
transformers>=4.41.1
open-clip-torch>=2.24.0
Pillow>=10.3.0
pillow-heif>=0.16.0
numpy>=1.26.4
opencv-python-headless>=4.9.0.80

# Storage
minio>=7.2.7

# Utilities
python-dotenv>=1.0.1
httpx>=0.27.0
aiofiles>=23.2.1

# Testing
pytest>=8.2.0
